Updated EDA, joins with all data
Some updated EDA for values in training or testing but not in
users.csv.

Left-joined data instead of just inner joined data. Note that the users
file had the NA values replaced with the mean of that column. Any NA’s
generated are from users who are not in the users.csv file.